requirements.txt
joblib==1.1.0
matplotlib==3.5.1
numpy==1.21.4
pandas==1.3.4
Pillow==9.0.0
plotly==5.5.0
seaborn==0.11.2
streamlit==1.2.0
sklearn
